* lisp/fringe.el (fringe-mode): Doc fix. * src/window.h (struct window): Change type of 'fringes_outside_margins' to bitfield.  Fix comment.  Adjust users accordingly. (struct window): Change type of 'window_end_bytepos' to ptrdiff_t. Adjust comment. * src/xdisp.c (try_window_id): Change type of 'first_vpos' and 'vpos' to ptrdiff_t.

+++ src/xdisp.c
@@ -17761,8 +17761,8 @@ try_window_id (struct window *w)
     {
       /* Displayed to end of window, but no line containing text was
 	 displayed.  Lines were deleted at the end of the window.  */
-      int first_vpos = WINDOW_WANTS_HEADER_LINE_P (w) ? 1 : 0;
-      int vpos = XFASTINT (w->window_end_vpos);
+      ptrdiff_t first_vpos = WINDOW_WANTS_HEADER_LINE_P (w) ? 1 : 0;
+      ptrdiff_t vpos = XFASTINT (w->window_end_vpos);
       struct glyph_row *current_row = current_matrix->rows + vpos;
       struct glyph_row *desired_row = desired_matrix->rows + vpos;
